+++ CHANGELOG.md
@@ -9,7 +9,9 @@ You can also check the [release page](https://github.com/visualize-admin/visuali
 
 ## Unreleased
 
-Nothing yet.
+- Fixes
+  - Conslidated behavior of setting initial filters (top-most hierarchy value) when filter was not present and multi-filter was removed
+  - Fixed switching between segmentation dimensions in column charts
 
 # [3.24.1] - 2023-11-13